requirements.txt
pyQt5
pillow